Refactor, add configs and dependency injection
