+++ utils/config.py
@@ -1,0 +1,26 @@
+from easydict import EasyDict
+from pathlib import Path
+import yaml
+
+def log_config_to_file(cfg, pre='cfg', logger=None):
+    for key, val in cfg.items():
+        if isinstance(cfg[key], EasyDict):
+            logger.info('\n%s.%s = edict()' % (pre, key))
+            log_config_to_file(cfg[key], pre=pre + '.' + key, logger=logger)
+            continue
+        logger.info('%s.%s: %s' % (pre, key, val))
+
+def cfg_from_yaml_file(cfg_file, config):
+    with open(cfg_file, 'r') as f:
+        try:
+            new_config = yaml.load(f, Loader=yaml.FullLoader)
+        except:
+            new_config = yaml.load(f)
+        config.update(EasyDict(new_config))
+
+    return config
+
+
+cfg = EasyDict()
+cfg.ROOT_DIR = (Path(__file__).resolve().parent / '../').resolve()
+
